Convert templates from ES6 template strings to Mustache templates.
We're making this change to improve the security of the extension. As it
stands, if api.munchdb.com was compromised, an attacker would be able to
infect whatever HTML they please into a user's browsing experience as
we are simply generating the HTML and then inserting it via the
innerHTML DOM method.

Now, we are using Hogan.js to precompile a set of HTML templates to
functions - which then render our content safely. For added protection
we are also parsing the rendered template string back into a DOM node
tree before appending to the live DOM.